Remove exception module
From Kraken Technologies handbook: "prefer to define exception types in
the same module as where they are raised."
